fixed tests, extracted 'context-dirtying' tests to own class
