+++ hawkbit-repository/hawkbit-repository-jpa/src/test/java/org/eclipse/hawkbit/repository/jpa/ControllerManagementTest.java
@@ -70,7 +70,6 @@ import org.eclipse.hawkbit.repository.model.TargetUpdateStatus;
 import org.eclipse.hawkbit.repository.test.matcher.Expect;
 import org.eclipse.hawkbit.repository.test.matcher.ExpectEvents;
 import org.eclipse.hawkbit.repository.test.util.WithSpringAuthorityRule;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -677,20 +676,6 @@ public class ControllerManagementTest extends AbstractJpaIntegrationTest {
 
         assertThat(result).hasSize(3);
         result.forEach((key, value) -> assertThat(value).hasSize(1));
-    }
-
-    @Test
-    @Ignore
-    @Description("Verify that controller registration does not result in a TargetPollEvent if feature is disabled")
-    @ExpectEvents({ @Expect(type = TargetCreatedEvent.class, count = 1),
-            @Expect(type = TargetPollEvent.class, count = 0) })
-    public void targetPollEventNotSendIfDisabled() {
-        try {
-            repositoryProperties.setPublishTargetPollEvent(false);
-            controllerManagement.findOrRegisterTargetIfItDoesNotExist("AA", LOCALHOST);
-        } finally {
-            repositoryProperties.setPublishTargetPollEvent(true);
-        }
     }
 
     @Test
