+++ hawkbit-repository/hawkbit-repository-jpa/src/test/resources/jpa-test.properties
@@ -31,8 +31,3 @@ spring.jpa.properties.eclipselink.logging.level=FINE
 spring.jpa.properties.eclipselink.logging.level.sql=FINE
 spring.jpa.properties.eclipselink.logging.parameters=true
 # Debug utility functions - END
-
-# Repository - START
-hawkbit.server.repository.eagerPollPersistence=false
-hawkbit.server.repository.pollPersistenceFlushTime=1000
-# Repository - END
