+++ hawkbit-repository/hawkbit-repository-test/src/main/java/org/eclipse/hawkbit/repository/test/util/TenantEventCounter.java
@@ -10,11 +10,11 @@ package org.eclipse.hawkbit.repository.test.util;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.stream.Collectors;
 
 import org.eclipse.hawkbit.repository.event.TenantAwareEvent;
 import org.springframework.context.ApplicationEvent;
@@ -26,43 +26,23 @@ public class TenantEventCounter implements ApplicationListener<ApplicationEvent>
     // will be persisted even between multiple beans construction/destruction
     private static final Map<String, Set<TenantAwareEvent>> TENANT_EVENTS_COUNT = new ConcurrentHashMap<>();
 
-    public Map<Class<? extends TenantAwareEvent>, Integer> getEventsCount(final String tenant) {
-        final Set<? extends TenantAwareEvent> events = TENANT_EVENTS_COUNT.getOrDefault(tenant, Collections.emptySet());
-        final Map<Class<? extends TenantAwareEvent>, Integer> eventsCount = new HashMap<>();
-        events.forEach(e -> eventsCount.merge(e.getClass(), 1, Integer::sum));
-        return eventsCount;
-    }
-
     @Override
     public void onApplicationEvent(final ApplicationEvent event) {
         if (event instanceof TenantAwareEvent) {
             assertThat(((TenantAwareEvent) event).getTenant()).isNotBlank();
 
-            final Set<TenantAwareEvent> eventsCount = TENANT_EVENTS_COUNT.getOrDefault(
-                    ((TenantAwareEvent) event).getTenant(), new HashSet<>());
-            eventsCount.add((TenantAwareEvent) event);
-            TENANT_EVENTS_COUNT.put(((TenantAwareEvent) event).getTenant(), eventsCount);
+            synchronized (TENANT_EVENTS_COUNT) {
+                final Set<TenantAwareEvent> eventsCount = TENANT_EVENTS_COUNT.getOrDefault(((TenantAwareEvent) event).getTenant(), new HashSet<>());
+                eventsCount.add((TenantAwareEvent) event);
+                TENANT_EVENTS_COUNT.put(((TenantAwareEvent) event).getTenant(), eventsCount);
+            }
         }
     }
-//    private static final Map<String, Map<Class<? extends TenantAwareEvent>, Integer>> TENANT_EVENTS_COUNT = new ConcurrentHashMap<>();
-//
-//    public void onApplicationEvent(final TenantAwareEvent event) {
-//        assertThat(event.getTenant()).isNotBlank();
-//
-//        final Map<Class<? extends TenantAwareEvent>, Integer> eventsCount = TENANT_EVENTS_COUNT.getOrDefault(
-//                event.getTenant(), new ConcurrentHashMap<>());
-//        eventsCount.merge(event.getClass(), 1, Integer::sum);
-//        TENANT_EVENTS_COUNT.put(event.getTenant(), eventsCount);
-//    }
-//
-//    public Map<Class<? extends TenantAwareEvent>, Integer> getEventsCount(final String tenant) {
-//        return TENANT_EVENTS_COUNT.getOrDefault(tenant, Collections.emptyMap());
-//    }
-//
-//    @Override
-//    public void onApplicationEvent(final ApplicationEvent event) {
-//        if (event instanceof  TenantAwareEvent) {
-//            onApplicationEvent((TenantAwareEvent) event);
-//        }
-//    }
+
+    public Map<Class<? extends TenantAwareEvent>, Integer> getEventsCount(final String tenant) {
+        final Set<? extends TenantAwareEvent> events = TENANT_EVENTS_COUNT.getOrDefault(tenant, Collections.emptySet());
+        synchronized (TENANT_EVENTS_COUNT) {
+            return events.stream().collect(Collectors.toMap(e -> e.getClass(), e -> 1, Integer::sum));
+        }
+    }
 }
